Add SongContentAdapter
