+++ app/build.gradle
@@ -57,6 +57,7 @@ dependencies {
     implementation 'androidx.constraintlayout:constraintlayout:2.0.2'
     implementation 'androidx.core:core-ktx:1.3.2'
     implementation 'androidx.fragment:fragment-ktx:1.3.0-beta01' // Prevent crash when adding content
+    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.3.0-beta01'
     implementation 'androidx.recyclerview:recyclerview-selection:1.0.0'
     implementation 'com.google.android.material:material:1.2.1' // content_main.xml ConstraintLayout
     kapt "androidx.room:room-compiler:$room_version"
